Move serialization of transform node instances to the configuration cache into the "user" codec, as these instances can be referenced from user code (indirectly), but may not necessarily be scheduled for execution and so not serialized by the "internal" codec.
This is really a work around for another problem. All project transform node instances should be scheduled, otherwise dependency edges are missing from the work graph. However it is currently possible to add project dependencies to a configuration after its dependencies have been visited and this can mean nodes can be referenced without being scheduled.
